+++ ChangeLog.md
@@ -1,3 +1,7 @@
+## 0.13.0
+
+* build-tool-overrides
+
 ## 0.12.0
 
 * New create-plan flags: `--add-package`, `--expect-test-failure` and `--expect-haddock-failure`